Framework
I added the variables into the method desclarations so they can be accepted once there is something in place to pass them in but I gave them default values so that the code will still run as intended if nothing is passed in. I also fixed a couple minor spelling and gramatical errors in comments and error messages.