Added priority preferences
I made a change to the preprocessing utils file to alter the way that the cost is calculated. The new value takes into account the user's priority settings, though uses them in a somewhat less elegant way. In the future, I intended to add the bikelane and roadtype settings and find a more logical way to calculate the cost using this data, but for now, here's something that will take priorities into account.